Подгруппы

+++ ApiVkProject/Controllers/UsersController.cs
@@ -23,7 +23,7 @@ namespace ApiVkProject.Controllers
         }
 
         // GET: Users/Details/5
-        public async Task<IActionResult> Details(int? id)
+        public async Task<IActionResult> Details(long? id)
         {
             if (id == null)
             {
